requirements.txt
pyyaml~=6.0
Pillow~=9.01
